Remove data from buffer when deserialization fails
When data deserialization failed in a remote channel receiver, the binary
data buffer was not cleared and thus the next call to the receive
function tried to deserialize the same data. This led to an
unrecoverable error condition on the channel.

This is fixed by removing data from the receive buffer if
deserialization failed.